Rename "repositories" namespace to "repository"
