+++ server/src/Service/GetUsersService.php
@@ -2,7 +2,7 @@
 
 namespace Cora\Service;
 
-use Cora\Repositories\UserRepository;
+use Cora\Repository\UserRepository;
 use Cora\Utils\Paginator;
 
 class GetUsersService {
